+++ crates/objc2/CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Make `#[name = ...]` in `define_class!` optional. If not specified, the macro
   will choose a suitable default that makes loading multiple versions of the
   same library possible.
+* Added conversion trait impls from `ProtocolObject<P>` to `AnyObject`.
 
 ## Fixed
 * Fixed undefined behaviour when calling `AnyObject::class` on invalid objects.